+++ Code/Effects/WorldEffects/ShuffleSfxEffect.cs
@@ -82,20 +82,28 @@ public sealed class ShuffleSfxEffect : CrowdControlEffect
 
     private SlotId OnPlaySfx(On_SoundPlayer.orig_Play orig, SoundPlayer self, ref SoundStyle style, Vector2? position, SoundUpdateCallback callback)
     {
-        // Play a random sfx for the attempted sfx
-        var hash = Math.Abs((style.Identifier ?? style.SoundPath).GetHashCode());
-        var original = style;
+        SoundStyle shuffled = default;
         try
         {
-            style = VanillaSfx[(hash + _seed) % VanillaSfx.Length];
-            return orig.Invoke(self, ref style, position, callback);
+            // Play a random sfx for the attempted sfx; do not modify the provided style EVER!
+            var hash = Math.Abs((style.Identifier ?? style.SoundPath).GetHashCode());
+            shuffled = VanillaSfx[(hash + _seed) % VanillaSfx.Length] with
+            {
+                Identifier = style.Identifier,
+                IsLooped = style.IsLooped,
+                MaxInstances = style.MaxInstances,
+                PlayOnlyIfFocused = style.PlayOnlyIfFocused,
+                SoundLimitBehavior = style.SoundLimitBehavior,
+                Type = style.Type,
+                Volume = style.Volume
+            };
+            return orig.Invoke(self, ref shuffled, position);
         }
         catch (AssetLoadException)
         {
             // "AssetLoadException: Asset could not be found: "Sounds\Zombie_131" (seems to be very rare)
-            TerrariaUtils.WriteDebug($"Failed to shuffle sfx due to an exception. {nameof(AssetLoadException)}: {original.SoundPath} -> {style.SoundPath} (seed: {_seed}).");
-            style = original;
-            return orig.Invoke(self, ref style, position, callback);
+            TerrariaUtils.WriteDebug($"Failed to shuffle sfx due to an exception. {nameof(AssetLoadException)}: {style.SoundPath} -> {shuffled.SoundPath} (seed: {_seed}).");
+            return orig.Invoke(self, ref style, position);
         }
     }