+++ CHANGELOG.md
@@ -19,8 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed "Money Boost" effect to only detect NPC strikes from the affected player.
 - Changed "Drunk Mode" effect to shuffle items' tooltips.
 - Changed "Spawn Critters" effect to no longer spawn Empress Butterfly.
+- Changed "Spawn Town NPC" effect to name the spawned NPC after the redeeming viewer.
 - Fixed "Clear/Explode/Shuffle Inventory" effects not failing when there are no items.
 - Fixed "Screen Shake" effect not pausing when game is paused.
+- Fixed "Shuffle Sfx" effect not finishing properly.
 
 ## [1.1.3] - 2023-06-10